+++ projects/demo-playwright/tests/kit/textarea/textarea.pw.spec.ts
@@ -1,7 +1,8 @@
 import {TuiDocumentationPagePO, tuiGoto} from '@demo-playwright/utils';
+import {DemoRoute} from '@demo/routes';
 import {expect, test} from '@playwright/test';
 
-test.describe('Textarea', () => {
+test.describe('TextareaLegacy', () => {
     test.use({viewport: {width: 400, height: 500}});
 
     test('correctly shows character with descenders inside placeholder', async ({
@@ -9,7 +10,7 @@ test.describe('Textarea', () => {
     }) => {
         const placeholder = 'qwertypgj_';
 
-        await tuiGoto(page, `components/textarea/API?placeholder=${placeholder}`);
+        await tuiGoto(page, `${DemoRoute.TextareaLegacy}/API?placeholder=${placeholder}`);
 
         const {apiPageExample} = new TuiDocumentationPagePO(page);
         const textarea = apiPageExample.getByPlaceholder(placeholder);
@@ -19,5 +20,39 @@ test.describe('Textarea', () => {
         await expect
             .soft(apiPageExample)
             .toHaveScreenshot('01-character-with-descenders-inside-placeholder.png');
+    });
+
+    ['m', 'l'].forEach((size) => {
+        test(`size of ${size}`, async ({page}) => {
+            await tuiGoto(
+                page,
+                `${DemoRoute.TextareaLegacy}/API?tuiTextfieldSize=${size}`,
+            );
+            const {apiPageExample} = new TuiDocumentationPagePO(page);
+
+            await expect(apiPageExample).toHaveScreenshot(
+                `textarea-tuiTextfieldSize-${size}.png`,
+            );
+        });
+    });
+
+    test('line break text', async ({page, browserName}) => {
+        // eslint-disable-next-line playwright/no-skipped-test
+        test.skip(browserName !== 'chromium', 'Font flaky');
+
+        await tuiGoto(page, `${DemoRoute.TextareaLegacy}/API`);
+        const {apiPageExample} = new TuiDocumentationPagePO(page);
+        const textAreaComponent = apiPageExample.locator('tui-textarea');
+        const textarea = apiPageExample.getByRole('textbox').first();
+
+        await textarea.fill('1\n2\n3\n4');
+
+        await expect(textAreaComponent).toHaveScreenshot('textarea-line-break.png');
+
+        await page.locator('.t-row input[tuiSwitch]').first().click();
+
+        await expect(textAreaComponent).toHaveScreenshot(
+            'textarea-line-break-disabled.png',
+        );
     });
 });
