+++ projects/demo/src/modules/components/textfield/examples/3/index.html
@@ -13,7 +13,7 @@
 </tui-textfield>
 
 <tui-textfield tuiChevron>
-    <label tuiLabel>Select from options</label>
+    <label tuiLabel>Select option</label>
     <select
         placeholder="I am placeholder"
         tuiTextfield
