+++ projects/demo/src/modules/components/textfield/examples/4/index.html
@@ -12,7 +12,7 @@
 </tui-textfield>
 
 <tui-textfield>
-    <label tuiLabel>Click icon to toggle type</label>
+    <label tuiLabel>Click icon to toggle</label>
     <input
         placeholder="I am placeholder"
         tuiTextfield
