+++ projects/experimental/components/input-phone-international/input-phone-international.component.ts
@@ -104,7 +104,6 @@ const NOT_FORM_CONTROL_SYMBOLS = /[^+\d]/g;
     host: {
         type: 'tel',
         ngSkipHydration: 'true',
-        '[attr.readonly]': 'readOnly() || null',
         '[attr.inputmode]': '!ios && open() ? "none" : null',
         '[disabled]': 'disabled()',
         '(blur)': 'onTouched()',
